Bump to Scala.js 0.6.0-M3
Neither Shapeless nor otest are compatible with the latest Scala.js
version. Therefore, minitest is now used instead. Functions which
depended on Shapeless were commented out and may be dropped entirely
in the future. Lenses turned out to have bad IDE support and error
messages are incomprehensible. An alternative is to use Var() for
fields which are meant to be mutable. This also results in better
performance as the whole object does not need to be reconstructed.